+++ packages/orq-rc/src/models/operations/listannotationqueues.ts
@@ -183,7 +183,7 @@ export const ListAnnotationQueuesData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-08-14T03:59:16.790Z",
+    "2026-08-14T13:41:24.623Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {