+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3894,7 +3894,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KZZ6SJJK100AHDK08X6FDJGZ"),
+  id: z.string().default("01M0083HRX4KYNGHZGJXC3BKB9"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),