+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -221,7 +221,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KZZ6SJVXV7NQQ0X92PG5VF73"),
+  _id: z.string().default("01M0083J42G2Z5PV5HMHJM9SGZ"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: UpdateDatasourceStatus$inboundSchema,